Add bulk delete stub

+++ src/AccountManagerDataAccess/src/main/java/org/cote/accountmanager/data/factory/DataFactory.java
@@ -346,6 +346,8 @@ public class DataFactory extends NameIdFactory {
 	public boolean deleteData(DataType data) throws FactoryException
 	{
 		removeFromCache(data);
+		if(bulkMode) return true;
+		
 		int deleted = deleteById(data.getId(), data.getOrganizationId());
 		/*
 		if (deleted > 0)
